requirements.txt
pygame==1.9.2
pyyaml==3.12
ws4py==0.4.2
